move Celled to grid.rs
